+++ core/src/network/p2p/event_loop.rs
@@ -451,9 +451,9 @@ impl EventLoop {
 					trace!("Identify Error event. PeerId: {peer_id:?}. Error: {error:?}");
 				},
 			},
-			SwarmEvent::Behaviour(BehaviourEvent::Mdns(event)) => match event {
-				mdns::Event::Discovered(addrs_list) => {
-					let peer_ids: Vec<PeerId> =
+			SwarmEvent::Behaviour(BehaviourEvent::Mdns(event)) => {
+				match event {
+					mdns::Event::Discovered(addrs_list) => {
 						addrs_list
 							.into_iter()
 							.filter(|(peer_id, multiaddr)| {
@@ -461,32 +461,17 @@ impl EventLoop {
 									.to_string()
 									.contains(Protocol::P2p(*peer_id).tag())
 							})
-							.map(|(peer_id, multiaddr)| {
+							.for_each(|(peer_id, multiaddr)| {
 								trace!("MDNS got peer with ID: {peer_id:#?} and Address: {multiaddr:#?}");
-								peer_id
-							})
-							.collect();
-
-					self.swarm.behaviour_mut().identify.push(peer_ids);
-				},
-				mdns::Event::Expired(addrs_list) => {
-					addrs_list.into_iter().for_each(|(peer_id, multiaddr)| {
-						trace!("MDNS got expired peer with ID: {peer_id:#?} and Address: {multiaddr:#?}");
-
-						if self
-							.swarm
-							.behaviour_mut()
-							.mdns
-							.discovered_nodes()
-							.any(|&p| p == peer_id)
-						{
-							self.swarm
-								.behaviour_mut()
-								.kademlia
-								.remove_address(&peer_id, &multiaddr);
-						}
-					});
-				},
+								if self.swarm.dial(multiaddr).is_err() {
+									warn!("Unable to dial peer with ID: {peer_id:#?}");
+								}
+							});
+					},
+					mdns::Event::Expired(addrs_list) => {
+						trace!("MDNS got expired peers: {:?}", addrs_list);
+					},
+				}
 			},
 			SwarmEvent::Behaviour(BehaviourEvent::AutoNat(event)) => match event {
 				autonat::Event::InboundProbe(e) => {